Changing namespace as requested on https://github.com/pear/Math_Stats/pull/2
